+++ R/zzz.R
@@ -9,13 +9,4 @@
   options(set6.unicode = NULL)
 }
 
-.onAttach <- function(libname, pkgname) {
-  packageStartupMessage("\n-----------------------------")
-  packageStartupMessage(
-    "\tset6 v", utils::packageVersion("set6"),
-    "\n\nGet started:\t?set6
-Changelog:\tset6News()"
-  )
-  packageStartupMessage("-----------------------------\n")
-}
 # nocov end